+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+ * `TransactionRecord.scheduleRef` - Reference to the scheduled transaction
+ * `TransactionReceipt.scheduledTransactionId`
+ * `ScheduleInfo.scheduledTransactionId`
+ * Feature to copy `TransactionId` of a transaction being scheduled
+   to the parent `ScheduleCreateTransaction` if one is set.
+
+### Fixed
+
+ * `TransactionId.toBytes()` should support `nonce` if set
+ * `TransactionId.fromBytes()` should support `nonce` if set
+
 ## v2.0.5-beta4
 
 ### Added